refactor: update extension imports to webapp/kernel/messages.ts
Point chrome-extension source and test files to the new canonical
location of the wire protocol types. Extension→webapp is the normal
dependency direction (already used for orchestrator, tools, etc.).

Part of #1443
